requirements.txt
google-generativeai>=0.1.0
